requirements.txt
GeoBases3K>=5.0.16
